First GUI screen - File Select
This is the GUI for the screen where the user selects the data category
(e.g. Publications, etc.) and browses to the CSV to select a specific
file.

The .cpp and .h are templates generated by Qt creator; I didn't do
anything to them. Not sure if those are needed, so I included them for
now to be safe.