Added a method to open VisualizationWindow from PresentationDashboardWindow
Added a method that looks for a doubleclick action in the QTreeWidget
and tests if it contains a faculty member (since we only graph by
faculty member). If it does, then open a new VisualizationWindow with
QList<PresentationRecord>, the time frame, and the faculty member name
as parameters for use in whatever actually creates the
VisualizationWindow.